move the refflat parser to its own file
